Test.Tasty.Ingredients.Basic

Description

This module exports the basic ingredients defined in the tasty packages.

Note that if defaultIngredients from Test.Tasty suits your needs, use that instead of importing this module.

Console test reporter

consoleTestReporter :: Ingredient

A simple console UI

newtype Quiet

Do not print test results (see README for details)

Constructors

Quiet Bool

newtype HideSuccesses

Report only failed tests

Constructors

HideSuccesses Bool

Listing tests

listingTests :: Ingredient

The ingredient that provides the test listing functionality

newtype ListTests

This option, when set to True, specifies that we should run in the «list tests» mode

Constructors

ListTests Bool

testsNames :: OptionSet -> TestTree -> [TestName]

Obtain the list of all tests in the suite

Adding options

includingOptions :: [OptionDescription] -> Ingredient

This ingredient doesn't do anything apart from registering additional options.

The option values can be accessed using askOption.